Move _get_package_summary from commands to common
It is used in separate commands, so will need to be imported into both.
