+++ src/pipx/commands/common.py
@@ -1,12 +1,16 @@
 import logging
+import shlex
 import shutil
 
 from pathlib import Path
 from shutil import which
 from typing import List
+from pipx import constants
+from pipx.colors import bold, red
 from pipx.emojies import hazard
 
 from pipx.util import WINDOWS, mkdir, rmdir
+from pipx.venv import Venv
 
 
 def expose_apps_globally(
@@ -76,3 +80,95 @@ def _symlink_package_apps(
                 f"{hazard}  Note: {app_name} was already on your PATH at "
                 f"{existing_executable_on_path}"
             )
+
+
+def _get_package_summary(
+    path: Path, *, package: str = None, new_install: bool = False
+) -> str:
+    venv = Venv(path)
+    python_path = venv.python_path.resolve()
+    if package is None:
+        package = path.name
+    if not python_path.is_file():
+        return f"   package {red(bold(package))} has invalid interpreter {str(python_path)}"
+
+    package_metadata = venv.package_metadata[package]
+
+    if package_metadata.package_version is None:
+        not_installed = red("is not installed")
+        return f"   package {bold(package)} {not_installed} in the venv {str(path)}"
+
+    apps = package_metadata.apps + package_metadata.apps_of_dependencies
+    exposed_app_paths = _get_exposed_app_paths_for_package(
+        venv.bin_path, apps, constants.LOCAL_BIN_DIR
+    )
+    exposed_binary_names = sorted(p.name for p in exposed_app_paths)
+    unavailable_binary_names = sorted(
+        set(package_metadata.apps) - set(exposed_binary_names)
+    )
+    # The following is to satisfy mypy that python_version is str and not
+    #   Optional[str]
+    python_version = (
+        venv.pipx_metadata.python_version
+        if venv.pipx_metadata.python_version is not None
+        else ""
+    )
+    return _get_list_output(
+        python_version,
+        python_path,
+        package_metadata.package_version,
+        package,
+        new_install,
+        exposed_binary_names,
+        unavailable_binary_names,
+    )
+
+
+def _get_exposed_app_paths_for_package(
+    venv_bin_path: Path, package_binary_names: List[str], local_bin_dir: Path
+):
+    bin_symlinks = set()
+    for b in local_bin_dir.iterdir():
+        try:
+            # sometimes symlinks can resolve to a file of a different name
+            # (in the case of ansible for example) so checking the resolved paths
+            # is not a reliable way to determine if the symlink exists.
+            # windows doesn't use symlinks, so the check is less strict.
+            if WINDOWS and b.name in package_binary_names:
+                is_same_file = True
+            else:
+                is_same_file = b.resolve().parent.samefile(venv_bin_path)
+            if is_same_file:
+                bin_symlinks.add(b)
+
+        except FileNotFoundError:
+            pass
+    return bin_symlinks
+
+
+def _get_list_output(
+    python_version: str,
+    python_path: Path,
+    package_version: str,
+    package: str,
+    new_install: bool,
+    exposed_binary_names: List[str],
+    unavailable_binary_names: List[str],
+) -> str:
+    output = []
+    output.append(
+        f"  {'installed' if new_install else ''} package {bold(shlex.quote(package))} {bold(package_version)}, {python_version}"
+    )
+
+    if not python_path.exists():
+        output.append(f"    associated python path {str(python_path)} does not exist!")
+
+    if new_install and exposed_binary_names:
+        output.append("  These apps are now globally available")
+    for name in exposed_binary_names:
+        output.append(f"    - {name}")
+    for name in unavailable_binary_names:
+        output.append(
+            f"    - {red(name)} (symlink missing or pointing to unexpected location)"
+        )
+    return "\n".join(output)
